// apps/tg-bot/src/app/modules/bot/constants/user-permission.enum.ts
//   allowMakeBan: false,

export enum UserPermissionEnum {
  IS_BASE_MODERATOR = 'IS_BASE_MODERATOR',
  ALLOW_PUBLISH_TO_CHANNEL = 'ALLOW_PUBLISH_TO_CHANNEL',
  ALLOW_DELETE_REJECTED_POST = 'ALLOW_DELETE_REJECTED_POST',
  ALLOW_RESTORE_DISCARDED_POST = 'ALLOW_RESTORE_DISCARDED_POST',
  ALLOW_SET_STRIKE = 'ALLOW_SET_STRIKE',
  ALLOW_MAKE_BAN = 'ALLOW_MAKE_BAN',
}